For consistency with other plotting functions, make it possible to plot to a file specifying the filename (and avoiding overwriting existing files)
