+++ openquake/baselib/general.py
@@ -1898,20 +1898,3 @@ class Deduplicate(Sequence):
 
     def __len__(self):
         return len(self.inv)
-
-
-def unique_filename(path):
-    """
-    Return a unique filename by appending a counter if needed.
-    Example: 'plot.png' -> 'plot_1.png' if 'plot.png' exists.
-    """
-    if not os.path.exists(path):
-        return path
-
-    root, ext = os.path.splitext(path)
-    counter = 1
-    new_path = f"{root}_{counter}{ext}"
-    while os.path.exists(new_path):
-        counter += 1
-        new_path = f"{root}_{counter}{ext}"
-    return new_path
